+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - MidiQoL Automation: Heroes Feast, Warding Bond.
 - DDB Importer Macro automation: Warding Bond.
+- Divine Smite used d6 damage, and corrections for 2014 feature.
 
 # 6.0.65